Rename test data to test_models.py to make it more understandable
Now it's clear that it's a test, unlike the more vague
`models_mysql.py` which people can easily mistake for an essential
part of the script.

(I'm not saying the models file is *not* essential: there's no point
in running the script if you don't have any data for it to work on.
But then again, this *particular* models file is probably not
essential. If anyone is willing to download this script, they'll
probably have some models of their own that they want to use).